Move cached & uncached serializers to their files
